Add basic support for windows
I wanted to keep it as simple as possible, so didn't even add guard
pages. I'll do it in a separate PR

I tried but just could not get it to work with gcc from mingw64, though
gcc from clang64 works, and clang from mingw64 and clang64 works.
Probably a compiler bug or something. I was getting a segfault when
calling coroutine_restore_context in coroutine_switch_context ..